requirements.txt
aamqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==5.2.5
django-admin-interface==0.30.1
django-allauth==65.11.1
django-appconf==1.1.0
django-celery-email==3.0.0
django-colorfield==0.14.0
django-crispy-forms==2.0
django-filter==25.1
djangorestframework==3.16.1
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
celery==5.3.6
redis==5.0.1